# cart/cart_item/router.py
from typing import Annotated
from sqlmodel import Session
from fastapi import APIRouter, Depends


from cart.cart_item import schema, crud, form
from lib.dependencies import ShopIDDep, LivemodeDep
from lib.session import get_session

router = APIRouter(prefix="/v1/carts")


@router.post("/{cart_id}/cart_items", response_model=schema.CartItem)
def create_cart_item(
    x_shop_id: ShopIDDep,
    x_livemode: LivemodeDep,
    cart_id: str,
    cart_item: Annotated[schema.CartItemCreate, Depends(form.create_cart_item_form)],
    db: Annotated[Session, Depends(get_session)],
):
    new_cart_item = crud.create_cart_item(
        x_shop_id,
        x_livemode,
        cart_id,
        cart_item,
        db,
    )
    return new_cart_item


@router.post(
    "/{cart_id}/cart_items/{cart_item_id}", response_model=schema.CartItemUpdate
)
def update_cart_item(
    x_shop_id: ShopIDDep,
    x_livemode: LivemodeDep,
    cart_id: str,
    cart_item_id: str,
    cart_item: Annotated[schema.CartItemUpdate, Depends(form.update_cart_item_form)],
    db: Annotated[Session, Depends(get_session)],
):
    updated_cart_item = crud.update_cart_item(
        x_shop_id,
        x_livemode,
        cart_id,
        cart_item_id,
        cart_item,
        db,
    )
    return updated_cart_item


@router.get("/{cart_id}/cart_items/{cart_item_id}", response_model=schema.CartItem)
def retrieve_cart_item(
    x_shop_id: ShopIDDep,
    x_livemode: LivemodeDep,
    cart_id: str,
    cart_item_id: str,
    db: Annotated[Session, Depends(get_session)],
):
    cart_item = crud.retrieve_cart_item(
        x_shop_id,
        x_livemode,
        cart_id,
        cart_item_id,
        db,
    )
    return cart_item


@router.get("/{cart_id}/cart_items", response_model=schema.CartItemList)
def list_cart_items(
    x_shop_id: ShopIDDep,
    x_livemode: LivemodeDep,
    cart_id: str,
    db: Annotated[Session, Depends(get_session)],
):
    # TODO skip limit
    cart_items_list = crud.list_cart_items(
        x_shop_id,
        x_livemode,
        cart_id,
        db,
    )
    return cart_items_list


@router.delete(
    "/{cart_id}/cart_items/{cart_item_id}", response_model=schema.CartItemDelete
)
def delete_cart_item(
    x_shop_id: ShopIDDep,
    x_livemode: LivemodeDep,
    cart_id: str,
    cart_item_id: str,
    db: Annotated[Session, Depends(get_session)],
):
    deleted_id = crud.delete_cart_item(
        x_shop_id,
        x_livemode,
        cart_id,
        cart_item_id,
        db,
    )
    deleted_cart_item = schema.CartItemDelete(
        id=cart_item_id,
        deleted=deleted_id is not None,
    )
    return deleted_cart_item