# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/v1/carts": {
      "get": {
        "summary": "List Carts",
        "operationId": "list_carts_v1_carts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartList" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Cart",
        "operationId": "create_cart_v1_carts_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_cart_v1_carts_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Cart" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/carts/{cart_id}": {
      "post": {
        "summary": "Update Cart",
        "operationId": "update_cart_v1_carts__cart_id__post",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Cart Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_cart_v1_carts__cart_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Cart" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve Cart",
        "operationId": "retrieve_cart_v1_carts__cart_id__get",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Cart Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Cart" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Cart",
        "operationId": "delete_cart_v1_carts__cart_id__delete",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Cart Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartDelete" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/carts/{cart_id}/cart_items": {
      "post": {
        "summary": "Create Cart Item",
        "operationId": "create_cart_item_v1_carts__cart_id__cart_items_post",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Cart Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_cart_item_v1_carts__cart_id__cart_items_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartItem" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Cart Items",
        "operationId": "list_cart_items_v1_carts__cart_id__cart_items_get",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Cart Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartItemList" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/carts/{cart_id}/cart_items/{cart_item_id}": {
      "post": {
        "summary": "Update Cart Item",
        "operationId": "update_cart_item_v1_carts__cart_id__cart_items__cart_item_id__post",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Cart Id" }
          },
          {
            "name": "cart_item_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Cart Item Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_cart_item_v1_carts__cart_id__cart_items__cart_item_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartItemUpdate" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve Cart Item",
        "operationId": "retrieve_cart_item_v1_carts__cart_id__cart_items__cart_item_id__get",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Cart Id" }
          },
          {
            "name": "cart_item_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Cart Item Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartItem" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Cart Item",
        "operationId": "delete_cart_item_v1_carts__cart_id__cart_items__cart_item_id__delete",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Cart Id" }
          },
          {
            "name": "cart_item_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Cart Item Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartItemDelete" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_create_cart_item_v1_carts__cart_id__cart_items_post": {
        "properties": {
          "cart": { "type": "string", "title": "Cart" },
          "price": { "type": "string", "title": "Price" },
          "quantity": { "type": "integer", "title": "Quantity", "default": 1 }
        },
        "type": "object",
        "required": ["cart", "price"],
        "title": "Body_create_cart_item_v1_carts__cart_id__cart_items_post"
      },
      "Body_create_cart_v1_carts_post": {
        "properties": {
          "price": { "type": "string", "title": "Price" },
          "quantity": { "type": "integer", "title": "Quantity", "default": 1 }
        },
        "type": "object",
        "required": ["price"],
        "title": "Body_create_cart_v1_carts_post"
      },
...
      "Body_update_cart_item_v1_carts__cart_id__cart_items__cart_item_id__post": {
        "properties": {
          "quantity": { "type": "integer", "title": "Quantity" }
        },
        "type": "object",
        "required": ["quantity"],
        "title": "Body_update_cart_item_v1_carts__cart_id__cart_items__cart_item_id__post"
      },
      "Body_update_cart_v1_carts__cart_id__post": {
        "properties": { "status": { "type": "string", "title": "Status" } },
        "type": "object",
        "required": ["status"],
        "title": "Body_update_cart_v1_carts__cart_id__post"
      },
...
      "Cart": {
        "properties": {
          "created": { "type": "integer", "title": "Created" },
          "updated": { "type": "integer", "title": "Updated" },
          "livemode": { "type": "boolean", "title": "Livemode" },
          "id": { "type": "string", "title": "Id" },
          "object": { "type": "string", "title": "Object", "default": "cart" },
          "status": { "$ref": "#/components/schemas/CartStatusEnum" },
          "cart_items": { "$ref": "#/components/schemas/CartItemList" }
        },
        "type": "object",
        "required": [
          "created",
          "updated",
          "livemode",
          "id",
          "status",
          "cart_items"
        ],
        "title": "Cart"
      },
      "CartDelete": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "object": { "type": "string", "title": "Object", "default": "cart" },
          "deleted": { "type": "boolean", "title": "Deleted" }
        },
        "type": "object",
        "required": ["id", "deleted"],
        "title": "CartDelete"
      },
      "CartItem": {
        "properties": {
          "created": { "type": "integer", "title": "Created" },
          "updated": { "type": "integer", "title": "Updated" },
          "livemode": { "type": "boolean", "title": "Livemode" },
          "quantity": { "type": "integer", "title": "Quantity" },
          "id": { "type": "string", "title": "Id" },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "cart_item"
          },
          "price": { "$ref": "#/components/schemas/Price" }
        },
        "type": "object",
        "required": [
          "created",
          "updated",
          "livemode",
          "quantity",
          "id",
          "price"
        ],
        "title": "CartItem"
      },
      "CartItemDelete": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "cart_item"
          },
          "deleted": { "type": "boolean", "title": "Deleted" }
        },
        "type": "object",
        "required": ["id", "deleted"],
        "title": "CartItemDelete"
      },
      "CartItemList": {
        "properties": {
          "object": { "type": "string", "title": "Object", "default": "list" },
          "url": { "type": "string", "title": "Url" },
          "has_more": { "type": "boolean", "title": "Has More" },
          "data": {
            "items": { "$ref": "#/components/schemas/CartItem" },
            "type": "array",
            "title": "Data",
            "default": []
          }
        },
        "type": "object",
        "required": ["url", "has_more"],
        "title": "CartItemList"
      },
      "CartItemUpdate": {
        "properties": {
          "quantity": { "type": "integer", "title": "Quantity" }
        },
        "type": "object",
        "required": ["quantity"],
        "title": "CartItemUpdate"
      },
      "CartList": {
        "properties": {
          "object": { "type": "string", "title": "Object", "default": "list" },
          "url": { "type": "string", "title": "Url", "default": "/v1/carts" },
          "has_more": { "type": "boolean", "title": "Has More" },
          "data": {
            "items": { "$ref": "#/components/schemas/Cart" },
            "type": "array",
            "title": "Data",
            "default": []
          }
        },
        "type": "object",
        "required": ["has_more"],
        "title": "CartList"
      },
      "CartStatusEnum": {
        "type": "string",
        "enum": ["requires_payment", "abandoned", "paid"],
        "title": "CartStatusEnum"
      },
...
      "CustomerUnitAmount": {
        "properties": {
          "maximum": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Maximum"
          },
          "minimum": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Minimum"
          },
          "preset": { "type": "integer", "title": "Preset", "default": 1 }
        },
        "type": "object",
        "title": "CustomerUnitAmount"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Price": {
        "properties": {
          "created": { "type": "integer", "title": "Created" },
          "updated": { "type": "integer", "title": "Updated" },
          "livemode": { "type": "boolean", "title": "Livemode" },
          "active": { "type": "boolean", "title": "Active" },
          "currency": { "type": "string", "title": "Currency" },
          "type": { "$ref": "#/components/schemas/PriceTypeEnum" },
          "unit_amount": { "type": "integer", "title": "Unit Amount" },
          "default": { "type": "boolean", "title": "Default" },
          "customer_unit_amount": {
            "anyOf": [
              { "$ref": "#/components/schemas/CustomerUnitAmount" },
              { "type": "null" }
            ]
          },
          "recurring": {
            "anyOf": [
              { "$ref": "#/components/schemas/Recurring" },
              { "type": "null" }
            ]
          },
          "id": { "type": "string", "title": "Id" },
          "object": { "type": "string", "title": "Object", "default": "price" }
        },
        "type": "object",
        "required": [
          "created",
          "updated",
          "livemode",
          "active",
          "currency",
          "type",
          "unit_amount",
          "default",
          "customer_unit_amount",
          "recurring",
          "id"
        ],
        "title": "Price"
      },
...
      "PriceTypeEnum": {
        "type": "string",
        "enum": ["one_time", "subscription"],
        "title": "PriceTypeEnum"
      },
...
      "Recurring": {
        "properties": {
          "interval": { "$ref": "#/components/schemas/RecurringTypeEnum" },
          "interval_count": { "type": "integer", "title": "Interval Count" }
        },
        "type": "object",
        "required": ["interval", "interval_count"],
        "title": "Recurring"
      },
      "RecurringTypeEnum": {
        "type": "string",
        "enum": ["day", "week", "month", "year"],
        "title": "RecurringTypeEnum"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}